+++ src/platform/linux/device.rs
@@ -365,7 +365,7 @@ impl AbstractDevice for Device {
         }
     }
 
-    fn mtu(&self) -> Result<i32> {
+    fn mtu(&self) -> Result<usize> {
         unsafe {
             let mut req = self.request();
 
@@ -373,14 +373,14 @@ impl AbstractDevice for Device {
                 return Err(io::Error::last_os_error().into());
             }
 
-            Ok(req.ifr_ifru.ifru_mtu)
+            Ok(req.ifr_ifru.ifru_mtu as usize)
         }
     }
 
-    fn set_mtu(&mut self, value: i32) -> Result<()> {
+    fn set_mtu(&mut self, value: usize) -> Result<()> {
         unsafe {
             let mut req = self.request();
-            req.ifr_ifru.ifru_mtu = value;
+            req.ifr_ifru.ifru_mtu = value as i32;
 
             if siocsifmtu(self.ctl.as_raw_fd(), &req) < 0 {
                 return Err(io::Error::last_os_error().into());
